Add SFX installer generator
The SFX part is based on a shell script prepended to a
tarball. To extract itself, the script cuts the shell
script part off and pipes the remainder to tar -xzf.

The SFX is then packed to a .uf2 file for ultimate
ease of installation. Dropping the .uf2 file to the
emulated drive installs NXT3 together with its files
and launches it.