fix(ui): fix startup animation speed
Previously, the startup animation was slower than on real NXTs.
The reason for this is that the animation used a loop counter
for timing. This forked fine on NXT as it's a hard realtime
system. On the other hand, EV3 currently does not maintain stable
1 ms loop times (the time is around 1-3 ms).

This commit fixes the animation by switching from loop counting
to real POSIX timers provided by the HAL layer.